Add files via upload
This file is the output of main.py. It takes on my laptop 6,8h to compute it. It's the first step of tetradic level analysis. It consists of 20536 formulas (with unknown places in the formulas). In the first element of each formula, there is the logical formula; In the second place are the four premisis (from which the logical formula is deduced).